user registerHandler for userSettingsChange
