+++ src/public/index.ts
@@ -13,7 +13,6 @@ export {
   TaskInfo,
   TeamInformation,
   FileOpenPreference,
-  UserSettingKeys,
 } from './interfaces';
 export {
   enablePrintCapability,
@@ -35,7 +34,6 @@ export {
   registerAppButtonHoverLeaveHandler,
   setFrameContext,
   shareDeepLink,
-  registerUserSettingsChangeHandler,
 } from './publicAPIs';
 export { returnFocus, navigateBack, navigateCrossDomain, navigateToTab } from './navigation';
 export { settings } from './settings';
