+++ src/public/interfaces.ts
@@ -167,19 +167,6 @@ export enum FileOpenPreference {
   Web = 'web',
 }
 
-/**
- * User setting changes that can be subscribed to,
- */
-export enum UserSettingKeys {
-  /**
-   * Use this key to subscribe to changes in user's file open preference
-   */
-  fileOpenPreference = 'fileOpenPreference',
-  /**
-   * Use this key to subscribe to theme changes
-   */
-  theme = 'theme',
-}
 export interface Context {
   /**
    * The Office 365 group ID for the team with which the content is associated.
