+++ src/public/publicAPIs.ts
@@ -9,7 +9,6 @@ import {
   Context,
   LoadContext,
   FrameContext,
-  UserSettingKeys,
 } from './interfaces';
 import { getGenericOnCompleteHandler } from '../internal/utils';
 import { logs } from '../private/logs';
@@ -328,18 +327,4 @@ export function initializeWithFrameContext(
 ): void {
   initialize(callback, validMessageOrigins);
   setFrameContext(frameContext);
-}
-
-/**
- * register a handler to be called when a user setting changes. The changed setting key & value is provided in the callback.
- * @param settingKeys List of user setting changes to subscribe
- * @param handler When a subscribed setting is updated this handler is called
- */
-export function registerUserSettingsChangeHandler(
-  settingKeys: UserSettingKeys[],
-  handler: (settingKey: UserSettingKeys, value: any) => void,
-): void {
-  ensureInitialized(FrameContexts.content);
-
-  Handlers.registerUserSettingsChangeHandler(settingKeys, handler);
 }
